Remove active/inactive handling from popup js
If the tab is on the Glimmer and Gloom page, then the side panel will be
enabled, so clicking on the action button will open the side panel, not
the popup.

The only time the popup is visible is if the side panel is not enabled,
so we can just display that message.