Remove WAITING_FOR_LEVEL_DATA state
This was kinda cool, but ultimately unnecessary. I used it at first to
keep track of the ackId from the request to generate the level and then
match that ackId against the response, but it doesn't really matter: if
we receive anything that looks like level data, we'll transition to the
COMPUTED_SOLUTION state without regard to whether to not we were
expecting it.

It also made it so that the UI would flash from COMPUTED_SOLUTION back
to WAITING_FOR_LEVEL_DATA whenever the user hit the restart button or
the play again button.